Rewire App Redux Added , Redux Actions and Reducer Creater.
