+++ fcv-front-end/.babelrc
@@ -1,0 +1,22 @@
+{
+    "presets": [
+        [
+            "@babel/env", 
+            {
+                "useBuiltIns": "entry",
+                "targets": {
+                    "browsers": [
+                        ">1%",
+                        "not dead",
+                        "not ie 11",
+                        "not op_mini all"
+                    ]
+                }
+            }
+        ],
+        "@babel/react"
+    ],
+    "plugins": [
+        "@babel/syntax-dynamic-import"
+    ]
+}
