+++ fcv-front-end/.gitignore
@@ -1,5 +1,3 @@
 node_modules
-package-lock.json
+package.json
 dist
-npm-debug.log
-.env
