+++ fcv-front-end/package.json
@@ -1,47 +1,44 @@
 {
   "name": "fcv-front-end",
-  "version": "1.0.0",
-  "description": "",
+  "version": "2.0.0",
+  "description": "Data Management front end interface of FCV",
   "main": "index.js",
   "scripts": {
     "start": "npm run watch",
-    "build": "webpack -p",
     "watch": "webpack --watch",
-    "build:prod": "webpack -p",
-    "stats": "webpack --profile --json > stats.json && webpack-bundle-analyzer stats.json dist"
+    "prod": "webpack -p",
+    "build": "webpack -b"
   },
-  "author": "Deep Panchal",
+  "author": "",
   "license": "MIT",
   "dependencies": {
-    "@material-ui/core": "^3.9.0",
+    "@material-ui/core": "^3.9.2",
     "@material-ui/icons": "^3.0.2",
-    "@material-ui/styles": "^3.0.0-alpha.9",
-    "dotenv": "^6.2.0",
-    "react": "^16.7.0-alpha.0",
-    "react-dom": "^16.7.0-alpha.0",
+    "@material-ui/styles": "^3.0.0-alpha.10",
+    "react": "^16.8.2",
+    "react-dom": "^16.8.2",
+    "react-redux": "^6.0.0",
     "react-router-dom": "^4.3.1",
+    "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0",
     "reset-css": "^4.0.1",
-    "typeface-roboto": "0.0.54",
     "validator": "^10.11.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.2.2",
+    "@babel/core": "^7.3.3",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.2.0",
     "@babel/polyfill": "^7.2.5",
     "@babel/preset-env": "^7.3.1",
-    "@babel/preset-react": "^7.0.0-beta.40",
+    "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.5",
-    "clean-webpack-plugin": "^1.0.1",
     "css-loader": "^2.1.0",
-    "html-webpack-plugin": "^4.0.0-beta.5",
-    "mini-css-extract-plugin": "^0.5.0",
+    "file-loader": "^3.0.1",
+    "html-webpack-plugin": "^3.2.0",
     "node-sass": "^4.11.0",
     "sass-loader": "^7.1.0",
-    "uglifyjs-webpack-plugin": "^2.1.1",
-    "webpack": "^4.29.0",
-    "webpack-bundle-analyzer": "^3.0.3",
-    "webpack-cli": "^3.2.1",
+    "style-loader": "^0.23.1",
+    "webpack": "^4.29.4",
+    "webpack-cli": "^3.2.3",
     "webpack-dev-server": "^3.1.14"
   }
 }
