+++ fcv-front-end/src/data/AppData.js
@@ -1,0 +1,4 @@
+export default {
+    // Back-end Request URL
+    requestURL: 'http://127.0.0.1:8000/api/',
+}
