+++ fcv-front-end/src/data/reducers/tableReducer.js
@@ -1,0 +1,14 @@
+const init = {
+    areas: false,
+}
+export default (state = init, action) => {
+    switch (action.type) {
+        case "CHANGE_TABLE":
+            state = {
+                ...state,
+                areas: action.areas
+            };
+            break;
+    }
+    return state
+}
