+++ fcv-front-end/src/styles/HeaderClasses.js
@@ -1,0 +1,80 @@
+import { fade } from '@material-ui/core/styles/colorManipulator';
+export default (theme) => ({
+    root: {
+      width: '100%',
+    },
+    grow: {
+      flexGrow: 1,
+    },
+    menuButton: {
+      marginLeft: -12,
+      marginRight: 20,
+    },
+    logo: {
+      width: '75px',
+      marginRight: '25px'
+    },
+    logoHide: {
+      display: 'none'
+    },
+    search: {
+      position: 'relative',
+      borderRadius: theme.shape.borderRadius,
+      backgroundColor: fade(theme.palette.common.white, 0.15),
+      '&:hover': {
+        backgroundColor: fade(theme.palette.common.white, 0.25),
+      },
+      marginLeft: 0,
+      width: 70,
+      [theme.breakpoints.up('sm')]: {
+        marginLeft: theme.spacing.unit,
+        width: 'auto',
+      },
+    },
+    searchIcon: {
+      width: theme.spacing.unit * 9,
+      height: '100%',
+      position: 'absolute',
+      pointerEvents: 'none',
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    inputRoot: {
+      color: 'inherit',
+      width: '100%',
+    },
+    menuItem: {
+      '&:hover': {
+        backgroundColor: theme.palette.primary.main,
+        '& $primary, & $icon': {
+          color: theme.palette.common.white,
+        },
+      },
+    },
+    primary: {},
+    icon: {},
+    inputInput: {
+      paddingTop: theme.spacing.unit,
+      paddingRight: theme.spacing.unit,
+      paddingBottom: theme.spacing.unit,
+      paddingLeft: theme.spacing.unit * 10,
+      transition: theme.transitions.create('width'),
+      width: '0',
+      '&:focus': {
+        width: window.innerWidth-180,
+      },
+      [theme.breakpoints.up('sm')]: {
+        width: 200,
+        '&:focus': {
+          width: 250,
+        },
+      },
+      [theme.breakpoints.up('md')]: {
+        width: 450,
+        '&:focus': {
+          width: 600,
+        },
+      },
+    },
+  });
